Move tests into src/tests/smartapp
