+++ Changelog
@@ -3,6 +3,7 @@ Version 0.7.2     unreleased
 	* Add Sphinx configuration path to .readthedocs.yml.
 	* Adjust GHA build process to allow builds for stacked PRs.
 	* Add a new `run clean` target to clean up generated data.
+	* Move unit tests from `tests` into `src/tests/smartapp`.
 	* Update the jinja2 transitive dependency to address CVE-2025-27516.
 	* Update the requests transitive dependency to address CVE-2024-47081.
 	* Update the urllib3 transitive dependency to address CVE-2025-50181.
