+++ pyproject.toml
@@ -11,7 +11,7 @@ include = [
    { path='LICENSE', format='sdist' },
    { path='README.md', format='sdist' },
    { path='docs', format='sdist' },
-   { path='tests', format='sdist' },
+   { path='src/tests', format='sdist' },
 ]
 classifiers = [
    "Operating System :: OS Independent",
@@ -122,7 +122,7 @@ pretty = true
 show_absolute_path = true
 show_column_numbers = true
 show_error_codes = true
-files = [ "src/smartapp", "tests" ]
+files = [ "src" ]
 check_untyped_defs = true
 disallow_any_generics = true
 disallow_incomplete_defs = true
