+++ src/tests/smartapp/fixtures/samples/request/CONFIGURATION-INITIALIZE.json
