+++ src/tests/smartapp/fixtures/samples/request/CONFIGURATION-PAGE.json
