+++ src/tests/smartapp/fixtures/samples/request/EVENT-DEVICE.json
