+++ src/tests/smartapp/fixtures/samples/response/CONFIGURATION-INITIALIZE.json
