+++ src/tests/smartapp/fixtures/samples/response/CONFIGURATION-PAGE-1of2.json
