+++ src/tests/smartapp/fixtures/samples/response/CONFIGURATION-PAGE-only.json
