+++ src/tests/smartapp/fixtures/samples/response/CONFIRMATION.json
