+++ src/tests/smartapp/fixtures/samples/response/INSTALL.json
