+++ src/tests/smartapp/fixtures/samples/response/OAUTH_CALLBACK.json
