+++ src/tests/smartapp/fixtures/samples/response/UNINSTALL.json
