+++ src/tests/smartapp/fixtures/samples/settings/BOOLEAN.json
