+++ src/tests/smartapp/fixtures/samples/settings/DECIMAL.json
