+++ src/tests/smartapp/fixtures/samples/settings/DEVICE.json
