+++ src/tests/smartapp/fixtures/samples/settings/EMAIL.json
