+++ src/tests/smartapp/fixtures/samples/settings/ENUM-GROUP.json
