+++ src/tests/smartapp/fixtures/samples/settings/ENUM.json
