+++ src/tests/smartapp/fixtures/samples/settings/ICON.json
