+++ src/tests/smartapp/fixtures/samples/settings/IMAGE.json
