+++ src/tests/smartapp/fixtures/samples/settings/PHONE.json
