+++ src/tests/smartapp/fixtures/samples/settings/TEXT.json
