+++ src/tests/smartapp/test_converter.py
@@ -11,7 +11,7 @@ from arrow import arrow
 
 from smartapp.converter import CONVERTER, deserialize_datetime, serialize_datetime
 from smartapp.interface import *
-from tests.testutil import load_dir
+from tests.smartapp.testutil import load_dir
 
 FIXTURE_DIR = os.path.join(os.path.dirname(__file__), "fixtures/samples")
 REQUEST_DIR = os.path.join(FIXTURE_DIR, "request")
